fix: change torrent_total_downloaded metric name
When a counter metric's name ends with '_total', prometheus_client
interprets that as the suffix for being a counter metric, and so
neglects to add its own, resulting in an incorrect final name. So, I've
swapped the words in the name so 'total' isn't at the end  now. It
matches with `total_size` better now anyway.